% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildWeightMatrix.R
\name{buildWeightMatrix}
\alias{buildWeightMatrix}
\title{Create a simulated weight matrix}
\usage{
buildWeightMatrix(nrow, ncol, prop = 0.5)
}
\arguments{
\item{nrow}{An integer value, the number of patients (in row)}

\item{ncol}{An integer value, the number of clusters (in columns)}

\item{prop}{An numerical value between 0 and 1, the proportion of non-zero coefficient
by row, by default 0.5}
}
\value{
A sparse \code{n} x \code{p} matrix
}
\description{
Create a simulated weight matrix
}
\examples{
W <- buildWeightMatrix(10,6)
}
