% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bxd}
\alias{bxd}
\title{Dataset of mice under chow and high fat dier}
\format{A list of three data frames with 64 rows (BXD data set)
\describe{
  \item{metabo}{Metabolomic quantification}
  \item{prot}{Protein abondance}
  \item{RNA}{Gene expression}
}}
\source{
\url{http://www.genenetwork.org/webqtl/main.py?FormID=sharinginfo&GN_AccessionId=714}
}
\description{
A list containing genomic data of BXD mice
}
