% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSelection.R
\name{compute_BIC}
\alias{compute_BIC}
\title{compute_BIC Compute BIC the model}
\usage{
compute_BIC(LL, pen = 1)
}
\arguments{
\item{LL}{result from computeLL function}

\item{pen}{penalty value in formula}
}
\description{
compute_BIC
}
