% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSelection.R
\name{compute_coph}
\alias{compute_coph}
\title{compute_coph the cophenetic correlation coefficient}
\usage{
compute_coph(res)
}
\arguments{
\item{res}{result from SolveInt function}
}
\value{
coph the cophenetic correlation coefficient
}
\description{
compute_coph
}
