% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSelection.R
\name{computeLL}
\alias{computeLL}
\title{computeLL Compute likelihood of the model}
\usage{
computeLL(Y, res)
}
\arguments{
\item{Y}{observations}

\item{res}{result from SolveInt function}
}
\description{
computeLL
}
