% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getW.R
\name{get.W}
\alias{get.W}
\title{get.W solve equation of W}
\usage{
get.W(Zbar, Ybar)
}
\arguments{
\item{Zbar}{summary of various blocks of ceofficient  matrices}

\item{Ybar}{summary of various block of  observations
library(Matrix)
library(dplyr)
W <- matrix(round(runif(20*2, 0, 5), 2), nrow=20, ncol=2)
W <- W/rowSums(W)
H1 <- as.matrix(rSpMatrix(2, 100000, nnz = 10000, rand.x= function(nnz) round(rnorm(nnz, 2, 0.2), 2) ))
H2 <- as.matrix(rSpMatrix(2, 1000, nnz = 100,  rand.x= function(nnz) round(rnorm(nnz, 2, 0.2), 2) ))
H3 <- as.matrix(rSpMatrix(2, 30000, nnz = 1000, rand.x= function(nnz) round(rnorm(nnz, 6, 0.2), 2) ))
Y1 <- as.matrix(W%*%H1 + rnorm(20, sd=1))
Y2 <- as.matrix(W%*%H2 + rnorm(20, sd=1))
Y3 <- as.matrix(W%*%H3 + rnorm(20, sd=1))
Ybar <- cbind(Y1, Y2, Y3)
Zbar <- cbind(H1, H2, H3) %>% t
We <- get.W(Zbar, Ybar)}
}
\description{
Get.W
}
