% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.W.supervised.R
\name{get.W.supervised}
\alias{get.W.supervised}
\title{get W in supervised way}
\usage{
get.W.supervised(Zbar, Ybar, group)
}
\arguments{
\item{Zbar}{summary of various blocks of ceofficient  matrices}

\item{Ybar}{summary of various block of  observations}

\item{group}{group for each individual}
}
\value{
Wc matrix W
}
\description{
get W in supervised way
}
