% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSNF.R
\name{init_SNF}
\alias{init_SNF}
\title{init_SNF}
\usage{
init_SNF(data, K, sigma = 0.5)
}
\arguments{
\item{data}{List of matrices.}

\item{K}{Number of clusters}

\item{sigma}{Variance for local model}
}
\value{
a list of \code{clust} the clustering of samples and
\code{fit} the results of the method SNF
}
\description{
init_SNF
}
