% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVE.R
\name{PVE}
\alias{PVE}
\title{PVE computing}
\usage{
PVE(data, H, W)
}
\arguments{
\item{data}{alist of original data of size $n\times J_k$}

\item{H}{a list of feature matrix $p\times J_k$}

\item{W}{a matrix of weights size $n\times{x}p$}
}
\value{
pve percentage of variance explained
}
\description{
PVE computing
}
