% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSpMatrix.R
\name{rSpMatrix}
\alias{rSpMatrix}
\title{rSpMatrix create a non negative matrix}
\usage{
rSpMatrix(nrow, ncol, nnz, rand.x = function(nnz) round(rnorm(nnz), 2))
}
\arguments{
\item{nrow}{number of row}

\item{ncol}{number of column}

\item{nnz}{sparsity coefficient}

\item{rand.x}{function to simulate data}
}
\description{
rSpMatrix random sparse matrix
}
\examples{
library(Matrix)
H <- rSpMatrix(2, 500, nnz = 500, rand.x= function(nnz) round(rnorm(nnz, 2, 0.2), 2) )
}
