% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SolveInt.R
\name{SolveInt}
\alias{SolveInt}
\title{SolveInt solve equation for H and W}
\usage{
SolveInt(
  Y,
  p,
  max.it = 20,
  flavor_mod = "glmnet",
  type = rep("none", length(Y)),
  init_flavor = "snf",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{Y}{observations, list of blocks of data}

\item{p}{The number of latent profiles}

\item{max.it}{by default 20. Maximum iteration of the algorithm, else until convergence
of matrix W}

\item{flavor_mod}{glmnet or (oem not implemented yet)}

\item{type}{type of data (by default all are "none"),
can take values into \code{ c("none", "methylation", "mutation")},
this argument allows to adapt modelisation by adding constraints.
For exemple methylation data takes values between 0 and 1, then H block corresponding to
methylation data takes values between 0 and 1.}

\item{init_flavor}{flavor to initialize W matrix (hclust, SNF, random, pca or svd)}

\item{verbose}{A logical value indicating whether to print extra information.
Defaults to FALSE}

\item{...}{other arguments}
}
\value{
A list that contains :
  \code{H} is a list of matrix (latent profiles for each block of data)
  \code{W} is a matrix (weight matrix).
  \code{loss} similarity of the clustering between two iterations
  \code{pve} Percentage of variance explained
}
\description{
SolveInt
}
