% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH.R
\name{get.H}
\alias{get.H}
\title{get.H solve equation for H}
\usage{
get.H(W, y, flavor_mod = "glmnet", verbose)
}
\arguments{
\item{W}{weight matrix}

\item{y}{observations}

\item{flavor_mod}{This refers to the mode of resolution of matrix H. The default is "glmnet", user can chose "oem" (not yet implemented).}

\item{verbose}{A logical value indicating whether to print extra information.
Defaults to FALSE}
}
\description{
Get.H
}
